requirements.txt
slackbot==0.4.1
PyYAML==3.12
requests==2.12.4